refactor: move sync logic into sync/ module
